infrastructure/lambda/file-processor/requirements.txt
cryptography>=41.0.0
PyJWT>=2.8.0
requests>=2.31.0
boto3>=1.34.0
